requirements.txt
appdirs==1.4.3
Django==1.9
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
six==1.10.0
appdirs==1.4.3
Django==1.9
django-kronos==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-crontab==2.1.1
python-dateutil==2.6.0
six==1.10.0
appdirs==1.4.3
Django==1.9
django-kronos==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-crontab==2.1.1
python-dateutil==2.6.0
selenium==3.3.3
six==1.10.0
appdirs==1.4.3
Django==1.9
django-kronos==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-crontab==2.1.1
python-dateutil==2.6.0
selenium==3.3.3
six==1.10.0
